feat: block parent scroll when popover is open
